updated redis flush command

+++ Script/Command/RedisFlushCommand.php
@@ -4,6 +4,8 @@ namespace Sludio\HelperBundle\Script\Command;
 
 use Predis\Client;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputDefinition;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -11,7 +13,9 @@ class RedisFlushCommand extends ContainerAwareCommand
 {
     protected function configure()
     {
-        $this->setName('sludio:redis:flush')->setDescription('FlushAll redis');
+        $this->setName('sludio:redis:flush')->setDefinition(new InputDefinition([
+            new InputArgument('clients', InputArgument::IS_ARRAY | InputArgument::OPTIONAL),
+        ]));
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
@@ -23,10 +27,21 @@ class RedisFlushCommand extends ContainerAwareCommand
             }
         }
 
-        foreach ($clients as $snc) {
-            $this->getContainer()->get($snc)->flushdb();
+        if (!empty($clients)) {
+            $allowed = $clients;
+            $clientsInput = $input->getArgument('clients');
+            if (!empty($clientsInput)) {
+                foreach ($clientsInput as &$client) {
+                    $client = 'snc_redis.'.$client;
+                }
+                $allowed = array_intersect($clients, $clientsInput);
+            }
+            foreach ($clients as $snc) {
+                if (in_array($snc, $allowed)) {
+                    $this->getContainer()->get($snc)->flushdb();
+                    $output->writeln('redis database '.$snc.' flushed');
+                }
+            }
         }
-
-        $output->writeln('redis database flushed');
     }
 }
